+++ Client/src/components/ProductPage.js
@@ -11,7 +11,7 @@ const ProductPage = () => {
       <Navbar />
       <div className="mx-auto w-full md:w-4/5 p-4 ">
         <h1 className="text-3xl font-semibold text-gray-900">
-          Product Details
+          商品详情
         </h1>
         {id && <ProductDetails productId={id} />}{" "}
         {/* Render the product details component */}