+++ Client/src/components/Profile/UserDetails.js
@@ -18,7 +18,7 @@ const UserDetails = ({
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
           <div className="mb-4">
             <UserField
-              label="Name"
+              label="姓名"
               name="fullName"
               value={editMode ? formData.fullName : userData.fullName}
               onChange={handleChange}
@@ -27,7 +27,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="Email"
+              label="邮箱"
               name="email"
               value={editMode ? formData.email : userData.email}
               onChange={handleChange}
@@ -36,7 +36,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="Phone No."
+              label="手机号"
               name="phoneNo"
               value={editMode ? formData.phoneNo : userData.phoneNo}
               onChange={handleChange}
@@ -45,7 +45,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="College"
+              label="学校"
               name="college"
               value={editMode ? formData.college : userData.college}
               onChange={handleChange}
@@ -54,7 +54,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="College ID"
+              label="学号"
               name="collegeId"
               value={editMode ? formData.collegeId : userData.collegeId}
               onChange={handleChange}
@@ -63,7 +63,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="Address"
+              label="地址"
               name="address"
               value={editMode ? formData.address : userData.address}
               onChange={handleChange}
@@ -72,7 +72,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="City"
+              label="城市"
               name="city"
               value={editMode ? formData.city : userData.city}
               onChange={handleChange}
@@ -81,7 +81,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="State"
+              label="省份"
               name="state"
               value={editMode ? formData.state : userData.state}
               onChange={handleChange}
@@ -90,7 +90,7 @@ const UserDetails = ({
           </div>
           <div className="mb-4">
             <UserField
-              label="Zip Code"
+              label="邮编"
               name="zipCode"
               value={editMode ? formData.zipCode : userData.zipCode}
               onChange={handleChange}
@@ -109,7 +109,7 @@ const UserDetails = ({
               onClick={handleSaveClick}
               className="w-full py-2 px-4 bg-yellow-500 text-white rounded-lg hover:bg-yellow-600 transition duration-300"
             >
-              Save
+              保存
             </button>
           ) : (
             // Read-only mode
@@ -118,7 +118,7 @@ const UserDetails = ({
               onClick={handleEditClick}
               className="w-full py-2 px-4 bg-yellow-500 text-white rounded-lg hover:bg-yellow-600 transition duration-300"
             >
-              Edit
+              编辑
             </button>
           )}
         </>